Renamed a few packages to internal in the OpenSprinkler binding. See openhab/static-code-analysis#100
