+++ addons/binding/org.openhab.binding.opensprinkler/OSGI-INF/OpenSprinklerDiscovery.xml
@@ -10,7 +10,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" modified="modified" name="org.openhab.binding.opensprinkler.discovery.OpenSprinklerDiscoveryService">
-   <implementation class="org.openhab.binding.opensprinkler.discovery.OpenSprinklerDiscoveryService"/>
+   <implementation class="org.openhab.binding.opensprinkler.internal.discovery.OpenSprinklerDiscoveryService"/>
    <service>
       <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
    </service>
