+++ addons/binding/org.openhab.binding.opensprinkler/src/main/java/org/openhab/binding/opensprinkler/internal/config/OpenSprinklerConfig.java
@@ -6,22 +6,32 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.opensprinkler.config;
+package org.openhab.binding.opensprinkler.internal.config;
 
 import static org.openhab.binding.opensprinkler.OpenSprinklerBindingConstants.DEFAULT_REFRESH_RATE;
-import static org.openhab.binding.opensprinkler.internal.api.OpenSprinklerApiConstants.DEFAULT_STATION_COUNT;
+import static org.openhab.binding.opensprinkler.internal.api.OpenSprinklerApiConstants.*;
 
 /**
- * The {@link OpenSprinklerPiConfig} class defines the configuration options
- * for the OpenSprinkler PI Thing.
+ * The {@link OpenSprinklerConfig} class defines the configuration options
+ * for the OpenSprinkler Thing.
  *
  * @author Chris Graham - Initial contribution
  */
-public class OpenSprinklerPiConfig {
+public class OpenSprinklerConfig {
     /**
-     * Number of stations to control.
+     * Hostname of the OpenSprinkler API.
      */
-    public int stations = DEFAULT_STATION_COUNT;
+    public String hostname = null;
+
+    /**
+     * The port the OpenSprinkler API is listening on.
+     */
+    public int port = DEFAULT_API_PORT;
+
+    /**
+     * The password to connect to the OpenSprinkler API.
+     */
+    public String password = DEFAULT_ADMIN_PASSWORD;
 
     /**
      * Number of seconds in between refreshes from the OpenSprinkler device.
