+++ addons/binding/org.openhab.binding.opensprinkler/src/main/java/org/openhab/binding/opensprinkler/internal/config/OpenSprinklerPiConfig.java
@@ -6,32 +6,22 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.opensprinkler.config;
+package org.openhab.binding.opensprinkler.internal.config;
 
 import static org.openhab.binding.opensprinkler.OpenSprinklerBindingConstants.DEFAULT_REFRESH_RATE;
-import static org.openhab.binding.opensprinkler.internal.api.OpenSprinklerApiConstants.*;
+import static org.openhab.binding.opensprinkler.internal.api.OpenSprinklerApiConstants.DEFAULT_STATION_COUNT;
 
 /**
- * The {@link OpenSprinklerConfig} class defines the configuration options
- * for the OpenSprinkler Thing.
+ * The {@link OpenSprinklerPiConfig} class defines the configuration options
+ * for the OpenSprinkler PI Thing.
  *
  * @author Chris Graham - Initial contribution
  */
-public class OpenSprinklerConfig {
+public class OpenSprinklerPiConfig {
     /**
-     * Hostname of the OpenSprinkler API.
+     * Number of stations to control.
      */
-    public String hostname = null;
-
-    /**
-     * The port the OpenSprinkler API is listening on.
-     */
-    public int port = DEFAULT_API_PORT;
-
-    /**
-     * The password to connect to the OpenSprinkler API.
-     */
-    public String password = DEFAULT_ADMIN_PASSWORD;
+    public int stations = DEFAULT_STATION_COUNT;
 
     /**
      * Number of seconds in between refreshes from the OpenSprinkler device.
